+++ backend/src/main/java/sg/backend/controller/FundingStoryController.java
@@ -11,12 +11,12 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
-import sg.backend.dto.request.FundingStoryRequestDto;
-import sg.backend.dto.response.GetProjectResponseDto;
+import sg.backend.dto.request.wirtefunding.FundingStoryRequestDto;
+import sg.backend.dto.response.writefunding.project.GetProjectResponseDto;
 import sg.backend.dto.response.ResponseDto;
-import sg.backend.dto.response.file.DeleteFileResponseDto;
-import sg.backend.dto.response.ModifyContentResponseDto;
-import sg.backend.dto.response.file.UploadImageResponseDto;
+import sg.backend.dto.response.writefunding.file.DeleteFileResponseDto;
+import sg.backend.dto.response.writefunding.ModifyContentResponseDto;
+import sg.backend.dto.response.writefunding.file.UploadImageResponseDto;
 import sg.backend.service.implement.FundingStoryService;
 
 @RestController
