--- backend/src/main/java/sg/backend/dto/request/InsertTagRequestDto.java
@@ -1,12 +1,0 @@
-package sg.backend.dto.request;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-public class InsertTagRequestDto {
-    private String tagName;
-}
