+++ backend/src/main/java/sg/backend/dto/request/wirtefunding/PolicyRefundRequestDto.java
@@ -1,0 +1,12 @@
+package sg.backend.dto.request.wirtefunding;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+public class PolicyRefundRequestDto {
+    private String refund_policy;
+}
