+++ backend/src/main/java/sg/backend/dto/request/wirtefunding/PolicyRewardRequestDto.java
@@ -1,4 +1,4 @@
-package sg.backend.dto.request;
+package sg.backend.dto.request.wirtefunding;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -7,6 +7,6 @@ import lombok.Setter;
 @Getter
 @Setter
 @NoArgsConstructor
-public class PolicyRefundRequestDto {
-    private String refund_policy;
+public class PolicyRewardRequestDto {
+    private String reward_info;
 }
