+++ backend/src/main/java/sg/backend/dto/response/writefunding/GetFundingMainResponseDto.java
@@ -1,10 +1,11 @@
-package sg.backend.dto.response;
+package sg.backend.dto.response.writefunding;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import sg.backend.common.ResponseCode;
 import sg.backend.common.ResponseMessage;
+import sg.backend.dto.response.ResponseDto;
 
 @Getter
 public class GetFundingMainResponseDto extends ResponseDto {
