+++ backend/src/main/java/sg/backend/dto/response/writefunding/file/UploadImageResponseDto.java
@@ -1,4 +1,4 @@
-package sg.backend.dto.response.file;
+package sg.backend.dto.response.writefunding.file;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
@@ -8,22 +8,25 @@ import sg.backend.common.ResponseMessage;
 import sg.backend.dto.response.ResponseDto;
 
 @Getter
-public class UploadInfoFileResponseDto extends ResponseDto {
+public class UploadImageResponseDto extends ResponseDto {
 
-    private String orginalName;
-    private String uuid_name;
     private String url;
+    private String uuid_name;
 
-    private UploadInfoFileResponseDto(String orginalName, String uuid_name, String url) {
+    private UploadImageResponseDto(String uuid_name, String url){
         super(ResponseCode.SUCCESS, ResponseMessage.SUCCESS);
-        this.orginalName = orginalName;
         this.uuid_name = uuid_name;
         this.url = url;
     }
 
-    public static ResponseEntity<UploadInfoFileResponseDto> success(String orginalName, String uuid_name, String url) {
-        UploadInfoFileResponseDto result = new UploadInfoFileResponseDto(orginalName, uuid_name, url);
+    public static ResponseEntity<UploadImageResponseDto> success(String uuid_name, String url) {
+        UploadImageResponseDto result = new UploadImageResponseDto(uuid_name, url);
         return ResponseEntity.status(HttpStatus.OK).body(result);
+    }
+
+    public static ResponseEntity<ResponseDto> unsupported_file() {
+        ResponseDto result = new ResponseDto(ResponseCode.UNSUPPORTED_MEDIA_TYPE, ResponseMessage.UNSUPPORTED_MEDIA_TYPE);
+        return ResponseEntity.status(HttpStatus.UNSUPPORTED_MEDIA_TYPE).body(result);
     }
 
     public static ResponseEntity<ResponseDto> not_existed_file(){
@@ -31,7 +34,7 @@ public class UploadInfoFileResponseDto extends ResponseDto {
         return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(result);
     }
 
-    public static ResponseEntity<ResponseDto> not_existed_post(){
+    public static ResponseEntity<ResponseDto> not_existed_post() {
         ResponseDto result = new ResponseDto(ResponseCode.NOT_EXISTED_BOARD, ResponseMessage.NOT_EXISTED_BOARD);
         return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(result);
     }
