+++ backend/src/main/java/sg/backend/dto/response/writefunding/project/GetProjectResponseDto.java
@@ -1,13 +1,14 @@
-package sg.backend.dto.response;
+package sg.backend.dto.response.writefunding.project;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import sg.backend.common.ResponseCode;
 import sg.backend.common.ResponseMessage;
+import sg.backend.dto.response.ResponseDto;
 
 @Getter
-public class GetProjectResponseDto extends ResponseDto{
+public class GetProjectResponseDto extends ResponseDto {
 
     private String title;
