+++ backend/src/main/java/sg/backend/dto/response/writefunding/reward/GetPolicyResponseDto.java
@@ -1,13 +1,14 @@
-package sg.backend.dto.response;
+package sg.backend.dto.response.writefunding.reward;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import sg.backend.common.ResponseCode;
 import sg.backend.common.ResponseMessage;
+import sg.backend.dto.response.ResponseDto;
 
 @Getter
-public class GetPolicyResponseDto extends ResponseDto{
+public class GetPolicyResponseDto extends ResponseDto {
     private String refund_policy;
     private String reward_info;
     private GetPolicyResponseDto(String refund, String reward) {
