+++ backend/src/main/java/sg/backend/dto/response/writefunding/reward/MakeRewardResponseDto.java
@@ -1,20 +1,23 @@
-package sg.backend.dto.response;
+package sg.backend.dto.response.writefunding.reward;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import sg.backend.common.ResponseCode;
 import sg.backend.common.ResponseMessage;
+import sg.backend.dto.response.ResponseDto;
 
 @Getter
-public class ModifyContentResponseDto extends ResponseDto {
+public class MakeRewardResponseDto extends ResponseDto {
+    private Long reward_id;
 
-    private ModifyContentResponseDto() {
+    private MakeRewardResponseDto(Long reward_id) {
         super(ResponseCode.SUCCESS, ResponseMessage.SUCCESS);
+        this.reward_id = reward_id;
     }
 
-    public static ResponseEntity<ModifyContentResponseDto> success() {
-        ModifyContentResponseDto result = new ModifyContentResponseDto();
+    public static ResponseEntity<MakeRewardResponseDto> success(Long reward_id){
+        MakeRewardResponseDto result = new MakeRewardResponseDto(reward_id);
         return ResponseEntity.status(HttpStatus.OK).body(result);
     }
 
@@ -23,9 +26,8 @@ public class ModifyContentResponseDto extends ResponseDto {
         return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(result);
     }
 
-    public static ResponseEntity<ResponseDto> not_permission(){
-        ResponseDto result = new ResponseDto(ResponseCode.NO_PERMISSTION, ResponseMessage.NO_PERMISSTION);
+    public static ResponseEntity<ResponseDto> not_existed_data(){
+        ResponseDto result = new ResponseDto(ResponseCode.NOT_EXISTED_DATA, ResponseMessage.NOT_EXISTED_DATA);
         return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(result);
     }
-
 }
