+++ backend/src/main/java/sg/backend/service/implement/FundingRewardService.java
@@ -4,10 +4,15 @@ import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import sg.backend.dto.request.MakeRewardRequestDto;
-import sg.backend.dto.request.PolicyRefundRequestDto;
-import sg.backend.dto.request.PolicyRewardRequestDto;
+import sg.backend.dto.request.wirtefunding.MakeRewardRequestDto;
+import sg.backend.dto.request.wirtefunding.PolicyRefundRequestDto;
+import sg.backend.dto.request.wirtefunding.PolicyRewardRequestDto;
 import sg.backend.dto.response.*;
+import sg.backend.dto.response.writefunding.DeleteDataResponseDto;
+import sg.backend.dto.response.writefunding.ModifyContentResponseDto;
+import sg.backend.dto.response.writefunding.reward.GetMRewardResponseDto;
+import sg.backend.dto.response.writefunding.reward.GetPolicyResponseDto;
+import sg.backend.dto.response.writefunding.reward.MakeRewardResponseDto;
 import sg.backend.entity.Funding;
 import sg.backend.entity.Reward;
 import sg.backend.entity.State;
